src/requirements.txt
PyPDF2~=2.10.5
gTTS~=2.2.3
moviepy==1.0.3
openai
Pillow~=11.2.1
python-dotenv
Werkzeug>=2.0.0,<3.0.0
Flask>=2.0.0,<3.0.0
pyttsx3
langchain
llama_index
pdfplumber
pypdf
faiss-cpu
langchain_community
langchain_openai
langchain_text_splitters
torch
sklearn
openpyxl
requests
rouge-score
nltk